+++ pysat/tests/test_files.py
@@ -146,14 +146,14 @@ class TestNoDataDir():
         """Runs before every method to create a clean testing setup."""
         self.temporary_file_list = False
         # store current pysat directory
-        self.saved_data_path = pysat.data_dir
+        self.saved_data_path = pysat.params['data_dirs']
 
-        pysat.data_dir = ''
+        pysat.params.data['data_dirs'] = []
         reload(pysat._files)
 
     def teardown(self):
         """Runs after every method to clean up previous testing."""
-        pysat.data_dir = self.saved_data_path
+        pysat.params.data['data_dirs'] = self.saved_data_path
         reload(pysat._files)
 
     def test_no_data_dir(self):
@@ -175,11 +175,11 @@ class TestBasics():
         self.stop = dt.datetime(2009, 12, 31)
 
         # store current pysat directory
-        self.data_path = pysat.data_dir
+        self.data_paths = pysat.params['data_dirs']
 
         # create temporary directory
         self.tempdir = tempfile.TemporaryDirectory()
-        pysat.utils.set_data_dir(self.tempdir.name, store=False)
+        pysat.params.data['data_dirs'] = [self.tempdir.name]
 
         self.testInst = \
             pysat.Instrument(inst_module=pysat.instruments.pysat_testing,
@@ -190,7 +190,7 @@ class TestBasics():
 
     def teardown(self):
         """Runs after every method to clean up previous testing."""
-        pysat.utils.set_data_dir(self.data_path, store=False)
+        pysat.params.data['data_dirs'] = self.data_paths
         self.tempdir.cleanup()
         del self.testInst, self.out, self.tempdir, self.start, self.stop
 
@@ -312,6 +312,7 @@ class TestBasics():
         # use from_os function to get pandas Series of files and dates
         files = pysat.Files.from_os(data_path=self.testInst.files.data_path,
                                     format_str=root_fname)
+        print(files)
         # check overall length
         assert len(files) == 5761
         # check specific dates
@@ -361,11 +362,13 @@ class TestBasics():
                      use_doy=False, root_fname=root_fname)
         # create the same range of dates
         dates = pysat.utils.time.create_date_range(start, stop, freq='100min')
+        print(dates, self.testInst.files.data_path)
         pysat.instruments.pysat_testing.list_files = \
             functools.partial(list_files, version=self.version)
         inst = pysat.Instrument(platform='pysat', name='testing',
                                 update_files=True)
         reload(pysat.instruments.pysat_testing)
+        print(inst.files.files, inst.files.data_path)
         assert (np.all(inst.files.files.index == dates))
 
 
@@ -384,10 +387,10 @@ class TestInstWithFiles():
     def setup(self):
         """Runs before every method to create a clean testing setup."""
         # store current pysat directory
-        self.data_path = pysat.data_dir
+        self.data_paths = pysat.params['data_dirs']
         # create temporary directory
         self.tempdir = tempfile.TemporaryDirectory()
-        pysat.utils.set_data_dir(self.tempdir.name, store=False)
+        pysat.params.data['data_dirs'] = [self.tempdir.name]
         # create testing directory
         create_dir(temporary_file_list=self.temporary_file_list)
 
@@ -434,7 +437,7 @@ class TestInstWithFiles():
                          clean_level='clean',
                          update_files=True,
                          temporary_file_list=self.temporary_file_list)
-        pysat.utils.set_data_dir(self.data_path, store=False)
+        pysat.params.data['data_dirs'] = self.data_paths
         self.tempdir.cleanup()
         del self.tempdir, self.start, self.stop, self.start2, self.stop2
 
@@ -591,10 +594,10 @@ class TestInstWithFilesNonStandard():
     def setup(self):
         """Runs before every method to create a clean testing setup."""
         # store current pysat directory
-        self.data_path = pysat.data_dir
+        self.data_paths = pysat.params['data_dirs']
         # create temporary directory
         self.tempdir = tempfile.TemporaryDirectory()
-        pysat.utils.set_data_dir(self.tempdir.name, store=False)
+        pysat.params.data['data_dirs'] = [self.tempdir.name]
 
         self.start = dt.datetime(2008, 1, 11)
         self.stop = dt.datetime(2008, 1, 15)
@@ -623,7 +626,7 @@ class TestInstWithFilesNonStandard():
                          clean_level='clean',
                          update_files=True,
                          temporary_file_list=self.temporary_file_list)
-        pysat.utils.set_data_dir(self.data_path, store=False)
+        pysat.params.data['data_dirs'] = self.data_paths
         del self.start, self.stop
 
     def test_files_non_standard_pysat_directory(self):
@@ -807,10 +810,10 @@ class TestFilesRaceCondition():
     def setup(self):
         """Runs before every method to create a clean testing setup."""
         # store current pysat directory
-        self.data_path = pysat.data_dir
+        self.data_paths = pysat.params['data_dirs']
         # create temporary directory
         self.tempdir = tempfile.TemporaryDirectory()
-        pysat.utils.set_data_dir(self.tempdir.name, store=False)
+        pysat.params.data['data_dirs'] = [self.tempdir.name]
         # create testing directory
         create_dir(temporary_file_list=self.temporary_file_list)
 
@@ -856,7 +859,7 @@ class TestFilesRaceCondition():
                          clean_level='clean',
                          update_files=True,
                          temporary_file_list=self.temporary_file_list)
-        pysat.utils.set_data_dir(self.data_path, store=False)
+        pysat.params.data['data_dirs'] = self.data_paths
 
 # TODO: This needs to be replaced or expanded based on the tests that
 # portalocker uses