+++ pysat/utils/_core.py
@@ -17,32 +17,47 @@ def set_data_dir(path=None, store=False):
 
     Parameters
     ----------
-    path : string
+    path : string or list-like of str
         valid path to directory pysat uses to look for data
     store : bool
         if True, store data directory for future runs
 
     """
 
-    # account for a user prefix in the path, such as ~
-    path = os.path.expanduser(path)
-    # account for the presence of $HOME or similar
-    path = os.path.expandvars(path)
 
-    if os.path.isdir(path):
+    print('Input path ', path)
+    paths = np.asarray(path)
+    if paths.shape == ():
+        paths = [paths.tolist()]
+    elif paths.shape[0] > 1:
+        paths = paths.squeeze().tolist()
+    elif paths.shape[0] == 1:
+        paths = paths.tolist()
+
+    # Account for a user prefix in the path, such as ~
+    paths = [os.path.expanduser(path) for path in paths]
+    # Account for the presence of $HOME or similar
+    paths = [os.path.expandvars(path) for path in paths]
+    # Ensure all paths are valid
+    paths_check = [os.path.isdir(path) for path in paths]
+    print('Processed path ', paths)
+    if np.all(paths_check):
         if store:
             estr = ''.join(('pysat has moved to a central structure for ',
                             'storing parameters to disk. Please switch to ',
                             '`pysat.params["data_dir"] = path` instead.'))
             raise RuntimeError(estr)
 
-        pysat.data_dir = path
+        # Assign updated and validated paths
+        pysat.params.data['data_dirs'] = paths
+        # Store information
+        pysat.params.store()
 
-        # Reload libraries if already present
-        if hasattr(pysat, '_files'):
-            pysat._files = importlib.reload(pysat._files)
-        if hasattr(pysat, '_instrument'):
-            pysat._instrument = importlib.reload(pysat._instrument)
+        # # Reload libraries if already present
+        # if hasattr(pysat, '_files'):
+        #     pysat._files = importlib.reload(pysat._files)
+        # if hasattr(pysat, '_instrument'):
+        #     pysat._instrument = importlib.reload(pysat._instrument)
     else:
         raise ValueError(' '.join(('Path {:s} does not lead to a valid',
                                    'directory.')).format(path))