+++ CHANGELOG.md
@@ -27,6 +27,12 @@ This project adheres to [Semantic Versioning](http://semver.org/).
   - Added support for 'cycle' in addition to 'version' and 'revision' for
     filename conventions.
   - Added testing utility functions.
+  - Added support for multiple pysat data directories
+  - Reorganized .pysat directory to store instrument information under 
+    .pysat/instruments and .pysat/instruments/archive
+  - Added pysat.params, a central location to store and modify pysat parameters.
+    Parameters are stored at .pysat/pysat_settings.json
+  - Updated default pysat directory organization template
 - Deprecations
   - Migraged instruments to pysatMadrigal, pysatNASA, pysatSpaceWeather,
     pysatIncubator, pysatModels, pysatCDAAC, and pysatMissions
@@ -59,6 +65,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
   - Changed `pysat.instruments.methods.general.list_files` kwarg
     `fake_monthly_files_from_daily` to `file_cadence`
   - Changed name of Instrument method `default` to `preprocess`
+  - Removed `pysat.data_dir`. Information now at `pysat.params['data_dirs']`.
 - Documentation
   - Added info on how to register new instruments
   - Fixed description of tag and inst_id behaviour in testing instruments
@@ -82,6 +89,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
   - Removed unused input arguments
   - Corrects Instrument today, yesterday, and tomorrow methods by implementing
     datetime.datetime.utcnow
+  - Ensured pysat.Meta instances are immutable at Instrument instantiation
+  - Removed weak reference back to Instrument within Files class
 - Maintenance
   - nose dependency removed from unit tests
   - Specify dtype for empty pandas.Series for forward compatibility